pyproject.toml
[tool.poetry]
name = "remind_me_api"
version = "0.1.0"
description = ""
authors = ["Jesse Brink <jbrink0925@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
python-decouple = "^3.4"
APScheduler = "^3.7.0"
pydantic = "^1.8.1"
httpx = "^0.17.1"
Jinja2 = "^2.11.3"
aiofiles = "^0.6.0"
python-dateutil = "^2.8.1"
SQLAlchemy = "1.3.22"
requests = "^2.25.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
